requirements.txt
Django==5.2.4
drf-yasg==1.21.10
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
django-allauth==65.10.0
django-filter==25.1
openapi-codec==1.3.2
simplejson==3.20.1
python-dotenv==1.1.1
cryptography==45.0.6
requests==2.32.4
django-modeltranslation==0.19.16
phonenumbers==9.0.10
phonenumberslite==9.0.10
pillow==11.3.0